+++ docs/CHANGELOG.md
@@ -5,5 +5,6 @@
 - Write empty geometry with the shapefile driver
 - Do not ask for SRID when geometry is empty
 - Add the name of the table in the FGB header
-- Fix null way id on OSM file
+- Fix null way id on OSM file- 
 - Use new sonatype auth
+- Add ST_SnapToSelf function